Validate dense policy bus inputs and split witness runtime
