+++ tools/bootnode/README.md
@@ -18,3 +18,4 @@ export DATABASE_URL='postgres://postgres:postgres@127.0.0.1:5432/bootnode'
 ## Production (HTTPS + ACME)
 
 Set `--domain` / `--acme-email` / `--acme-cache-dir`, and bind to `:443`.
+
