+++ tools/bootnode/docker-compose.yml
@@ -39,3 +39,4 @@ services:
 volumes:
   bootnode_pg:
   bootnode_acme:
+
