+++ tools/bootnode/systemd/bootnode.env.example
@@ -23,3 +23,4 @@ BOOTNODE_OTEL_ENABLED=false
 # BOOTNODE_OTEL_OTLP_ENDPOINT=http://otel-collector:4317
 # BOOTNODE_OTEL_SERVICE_NAME=poolstellar-bootnode
 # BOOTNODE_OTEL_SAMPLE_RATIO=0.05
+
